Implement solution for Day 2, part 1
This adds a simple method to check if a given string is valid for the
requirements stated in the problem. We then filter the stream of lines
from the input file through this method to get all "valid" lines and
return the total count.

It's pretty obvious at this point there is going to be repeated code for
both the solutions and the tests that should be cleaned up. The common
boilerplate will be pulled out in a follow-up commit.