Make extra test case valid for part 2
I added this last test case to deal with an edge case for part one, but
the number range at the beginning is not valid for part 2. The second
index is out of bounds for the word. This could be handled in code, but
all the supplied inputs are valid so we can write the method assuming
valid input.