Implement solution to Day 3, part 2
There really wasn't any work to do here. The traverse() method could
easily be reused to implement the second part; we just had to call it
for each provided "slope".

The one change that had to be made was to have part 2 return a
BigInteger and stream reduce on BigIntegers since the final value
overflowed the int return type.