Use graphs to solve Day 7 part 1
This is a very simplistic graph implementation, but it's a convenient
way to represent the problem. We simply read the file, generate a graph
of bag objects, and count all the bags from which we can find a path to
the desired bag.

This makes some assumptions about the input. Clearly the graph is
directed, but it assumes there are no cycles or the program would loop
infinitely. This seemed like a reasonable assumption since a cycle
implies a bag contains itself, which is nonsense.